Changes all items melee and throw force to /tg/ values and bullet damage values. Renames OxygenLevel to OxygenDamage and it changes its behaviour, now it'll act alike brute or burn damage, if the sume of the damage types is over 100 you'll go to crit. Breathing reduces the OxygenDamage slowly, and at only 50 will make your character pass out to soft crit, and under 100 total dmg your character won't try to breath and instead will increase the OxygenDamage slowly. Removes the brain damage kill behaviour. Removes the unused ProjectileDamage variable from weapons, the damage value comes from the bullets themselves. Bleeding is limb-based and will now only trigger at over 40 brute damage and will heal once the limb that is bleeding is healed. Adds Attack() for p2p interactions, currently only used by bruisepacks and ointments. Adds bruisepacks and ointments pickuptriggers, using Attack() to heal on the targetted limb. Now the only health update that the players will receive will be the their own, the conscious state of other players is the exception. Separates conscious state away from HealthOverallMessage to a new HealthConsciousMessage.

+++ UnityProject/Assets/Scripts/UI/UI_HeartMonitor.cs
@@ -26,7 +26,7 @@ public class UI_HeartMonitor : MonoBehaviour
 	private int spriteStart;
 	private bool startMonitoring;
 	private float timeWait;
-	private int overallHealthCache = 100;
+	private float overallHealthCache = 100;
 
 	private void Start()
 	{
